requirements.txt
pandas==2.*
python-dotenv==1.*
supabase==2.*